#73 refactor: 내가 만든 차량 목록 컴포넌트 분리
